pyproject.toml
[project]
name = "car-edit-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "diffusers>=0.35.1",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "safetensors>=0.6.2",
    "torch>=2.8.0",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipython>=9.5.0",
    "jupyter>=1.1.1",
    "pytest>=8.4.1",
]
